Transfer home logic to HomeContainer, HomeTab only contains the skeleton
